+++ Example/Mediator/Test/Test/Classes/Private/Component/TestStatisticalViewController.m
@@ -198,7 +198,7 @@ FWPropertyWeak(FWTextTagCollectionView *, tagCollectionView);
 {
     self.collectionView.hidden = YES;
     FWWeakifySelf();
-    [self fwSetRightBarItem:@(UIBarButtonSystemItemRefresh) block:^(id  _Nonnull sender) {
+    [self fwSetRightBarItem:FWIcon.refreshImage block:^(id  _Nonnull sender) {
         FWStrongifySelf();
         if (self.collectionView.hidden) {
             self.collectionView.hidden = NO;