+++ FWFramework/Classes/Component/UIKit/Extension/UITextView+FWPlaceholder.m
@@ -38,27 +38,24 @@
         label.numberOfLines = 0;
         label.userInteractionEnabled = NO;
         label.font = self.font;
-        
         objc_setAssociatedObject(self, @selector(fwPlaceholderLabel), label, OBJC_ASSOCIATION_RETAIN_NONATOMIC);
-        
         [self fwUpdatePlaceholder];
+        [self insertSubview:label atIndex:0];
         
         // 监听当前输入框的文本改变通知
-        [self fwObserveNotification:UITextViewTextDidChangeNotification object:self target:self action:@selector(fwUpdatePlaceholder)];
+        [self fwObserveNotification:UITextViewTextDidChangeNotification object:self target:self action:@selector(fwTextDidChange)];
         
         // 监听当前输入框属性改变
-        [self fwObserveProperty:@"attributedText" target:self action:@selector(fwUpdatePlaceholder)];
+        [self fwObserveProperty:@"attributedText" target:self action:@selector(fwTextDidChange)];
+        [self fwObserveProperty:@"text" target:self action:@selector(fwTextDidChange)];
         [self fwObserveProperty:@"bounds" target:self action:@selector(fwUpdatePlaceholder)];
         [self fwObserveProperty:@"frame" target:self action:@selector(fwUpdatePlaceholder)];
-        [self fwObserveProperty:@"text" target:self action:@selector(fwUpdatePlaceholder)];
         [self fwObserveProperty:@"textAlignment" target:self action:@selector(fwUpdatePlaceholder)];
         [self fwObserveProperty:@"textContainerInset" target:self action:@selector(fwUpdatePlaceholder)];
         
         // 监听字体改变
         [self fwObserveProperty:@"font" block:^(UITextView *textView, NSDictionary *change) {
-            if (change[NSKeyValueChangeNewKey] != nil) {
-                textView.fwPlaceholderLabel.font = textView.font;
-            }
+            if (change[NSKeyValueChangeNewKey] != nil) textView.fwPlaceholderLabel.font = textView.font;
             [textView fwUpdatePlaceholder];
         }];
     }
@@ -68,21 +65,56 @@
 - (void)fwUpdatePlaceholder
 {
     if (self.text.length) {
-        [self.fwPlaceholderLabel removeFromSuperview];
+        self.fwPlaceholderLabel.hidden = YES;
         return;
     }
     
-    [self insertSubview:self.fwPlaceholderLabel atIndex:0];
-    
-    // iOS7以上
+    self.fwPlaceholderLabel.hidden = NO;
     self.fwPlaceholderLabel.textAlignment = self.textAlignment;
-    CGFloat lineFragmentPadding = self.textContainer.lineFragmentPadding;
-    UIEdgeInsets textContainerInset = self.textContainerInset;
-    CGFloat x = lineFragmentPadding + textContainerInset.left;
-    CGFloat y = textContainerInset.top;
-    CGFloat width = CGRectGetWidth(self.bounds) - x - lineFragmentPadding - textContainerInset.right;
+    CGFloat x = self.textContainer.lineFragmentPadding + self.textContainerInset.left;
+    CGFloat y = self.textContainerInset.top;
+    if (self.fwPlaceholderCursorCenter) {
+        y = self.fwPlaceholderInsetTop;
+    }
+    CGFloat width = CGRectGetWidth(self.bounds) - x - self.textContainer.lineFragmentPadding - self.textContainerInset.right;
     CGFloat height = [self.fwPlaceholderLabel sizeThatFits:CGSizeMake(width, 0)].height;
     self.fwPlaceholderLabel.frame = CGRectMake(x, y, width, height);
+}
+
+- (void)fwTextDidChange
+{
+    [self fwUpdatePlaceholder];
+    
+    if (self.fwPlaceholderCursorCenter) {
+        NSInteger currentHeight = ceil([self sizeThatFits:CGSizeMake(self.bounds.size.width, CGFLOAT_MAX)].height);
+        NSInteger placeholderHeight = ceil(self.fwPlaceholderHeight);
+        CGFloat insetTop = self.fwPlaceholderInsetTop;
+        if (placeholderHeight >= currentHeight) {
+            CGFloat textHeight = currentHeight - self.textContainerInset.top - self.textContainerInset.bottom;
+            insetTop = (placeholderHeight - textHeight) / 2.0;
+        }
+        UIEdgeInsets textContainerInset = self.textContainerInset;
+        textContainerInset.top = insetTop;
+        self.textContainerInset = textContainerInset;
+    }
+    
+    if (self.fwAutoHeightEnabled) {
+        NSInteger currentHeight = ceil([self sizeThatFits:CGSizeMake(self.bounds.size.width, CGFLOAT_MAX)].height);
+        // 如果显示placeholder，同时计算placeholder高度，取其中较大值
+        if (!self.fwPlaceholderLabel.isHidden) {
+            currentHeight = MAX(currentHeight, ceil(self.fwPlaceholderHeight));
+        }
+        currentHeight = MAX(self.fwMinHeight, MIN(currentHeight, self.fwMaxHeight));
+        if (currentHeight == self.fwLastHeight) return;
+        
+        CGRect frame = self.frame;
+        frame.size.height = currentHeight;
+        self.frame = frame;
+        if (self.fwHeightDidChange) {
+            self.fwHeightDidChange(currentHeight);
+        }
+        self.fwLastHeight = currentHeight;
+    }
 }
 
 #pragma mark - Public
@@ -95,7 +127,7 @@
 - (void)setFwPlaceholder:(NSString *)fwPlaceholder
 {
     self.fwPlaceholderLabel.text = fwPlaceholder;
-    [self fwUpdatePlaceholder];
+    [self fwTextDidChange];
 }
 
 - (NSAttributedString *)fwAttributedPlaceholder
@@ -106,7 +138,7 @@
 - (void)setFwAttributedPlaceholder:(NSAttributedString *)fwAttributedPlaceholder
 {
     self.fwPlaceholderLabel.attributedText = fwAttributedPlaceholder;
-    [self fwUpdatePlaceholder];
+    [self fwTextDidChange];
 }
 
 - (UIColor *)fwPlaceholderColor
@@ -117,6 +149,90 @@
 - (void)setFwPlaceholderColor:(UIColor *)fwPlaceholderColor
 {
     self.fwPlaceholderLabel.textColor = fwPlaceholderColor;
+}
+
+- (CGFloat)fwPlaceholderHeight
+{
+    return CGRectGetMaxY(self.fwPlaceholderLabel.frame) + self.textContainerInset.bottom;
+}
+
+- (BOOL)fwPlaceholderCursorCenter
+{
+    return [objc_getAssociatedObject(self, @selector(fwPlaceholderCursorCenter)) boolValue];
+}
+
+- (void)setFwPlaceholderCursorCenter:(BOOL)center
+{
+    objc_setAssociatedObject(self, @selector(fwPlaceholderCursorCenter), @(center), OBJC_ASSOCIATION_RETAIN_NONATOMIC);
+    objc_setAssociatedObject(self, @selector(fwPlaceholderInsetTop), center ? @(self.textContainerInset.top) : nil, OBJC_ASSOCIATION_RETAIN_NONATOMIC);
+    [self fwTextDidChange];
+}
+
+- (CGFloat)fwPlaceholderInsetTop
+{
+    NSNumber *value = objc_getAssociatedObject(self, @selector(fwPlaceholderInsetTop));
+    return value ? value.doubleValue : self.textContainerInset.top;
+}
+
+#pragma mark - AutoHeight
+
+- (BOOL)fwAutoHeightEnabled
+{
+    return [objc_getAssociatedObject(self, @selector(fwAutoHeightEnabled)) boolValue];
+}
+
+- (void)setFwAutoHeightEnabled:(BOOL)enabled
+{
+    objc_setAssociatedObject(self, @selector(fwAutoHeightEnabled), @(enabled), OBJC_ASSOCIATION_RETAIN_NONATOMIC);
+    [self fwTextDidChange];
+}
+
+- (CGFloat)fwMaxHeight
+{
+    NSNumber *value = objc_getAssociatedObject(self, @selector(fwMaxHeight));
+    return value ? value.doubleValue : CGFLOAT_MAX;
+}
+
+- (void)setFwMaxHeight:(CGFloat)height
+{
+    objc_setAssociatedObject(self, @selector(fwMaxHeight), @(height), OBJC_ASSOCIATION_RETAIN_NONATOMIC);
+}
+
+- (CGFloat)fwMinHeight
+{
+    return [objc_getAssociatedObject(self, @selector(fwMinHeight)) doubleValue];
+}
+
+- (void)setFwMinHeight:(CGFloat)height
+{
+    objc_setAssociatedObject(self, @selector(fwMinHeight), @(height), OBJC_ASSOCIATION_RETAIN_NONATOMIC);
+}
+
+- (void (^)(CGFloat))fwHeightDidChange
+{
+    return objc_getAssociatedObject(self, @selector(fwHeightDidChange));
+}
+
+- (void)setFwHeightDidChange:(void (^)(CGFloat))block
+{
+    objc_setAssociatedObject(self, @selector(fwHeightDidChange), block, OBJC_ASSOCIATION_COPY_NONATOMIC);
+}
+
+- (CGFloat)fwLastHeight
+{
+    return [objc_getAssociatedObject(self, @selector(fwLastHeight)) doubleValue];
+}
+
+- (void)setFwLastHeight:(CGFloat)height
+{
+    objc_setAssociatedObject(self, @selector(fwLastHeight), @(height), OBJC_ASSOCIATION_RETAIN_NONATOMIC);
+}
+
+- (void)fwAutoHeightWithMaxHeight:(CGFloat)maxHeight didChange:(void (^)(CGFloat))didChange
+{
+    self.fwMaxHeight = maxHeight;
+    if (didChange) self.fwHeightDidChange = didChange;
+    self.fwAutoHeightEnabled = YES;
 }
 
 @end